Fixed create new optimistic
