+++ packages/twenty-front/src/modules/action-menu/actions/record-actions/no-selection/components/CreateNewIndexRecordNoSelectionRecordAction.tsx
@@ -2,7 +2,7 @@ import { Action } from '@/action-menu/actions/components/Action';
 import { useContextStoreObjectMetadataItemOrThrow } from '@/context-store/hooks/useContextStoreObjectMetadataItemOrThrow';
 import { useCreateNewIndexRecord } from '@/object-record/record-table/hooks/useCreateNewIndexRecord';
 
-export const CreateNewTableRecordNoSelectionRecordAction = () => {
+export const CreateNewIndexRecordNoSelectionRecordAction = () => {
   const { objectMetadataItem } = useContextStoreObjectMetadataItemOrThrow();
 
   const { createNewIndexRecord } = useCreateNewIndexRecord({
