+++ packages/twenty-front/src/modules/object-record/record-board/record-board-column/components/RecordBoardColumnHeader.tsx
@@ -9,21 +9,14 @@ import { RecordBoardColumnHeaderAggregateDropdown } from '@/object-record/record
 import { RECORD_BOARD_COLUMN_WIDTH } from '@/object-record/record-board/constants/RecordBoardColumnWidth';
 import { RecordBoardColumnContext } from '@/object-record/record-board/record-board-column/contexts/RecordBoardColumnContext';
 import { hasAnySoftDeleteFilterOnViewComponentSelector } from '@/object-record/record-filter/states/hasAnySoftDeleteFilterOnView';
-import { recordGroupDefinitionsComponentSelector } from '@/object-record/record-group/states/selectors/recordGroupDefinitionsComponentSelector';
 import { RecordGroupDefinitionType } from '@/object-record/record-group/types/RecordGroupDefinition';
 import { recordIndexAggregateDisplayLabelComponentState } from '@/object-record/record-index/states/recordIndexAggregateDisplayLabelComponentState';
 import { recordIndexAggregateDisplayValueForGroupValueComponentFamilyState } from '@/object-record/record-index/states/recordIndexAggregateDisplayValueForGroupValueComponentFamilyState';
-import { recordIndexRecordIdsByGroupComponentFamilyState } from '@/object-record/record-index/states/recordIndexRecordIdsByGroupComponentFamilyState';
-import { useUpsertRecordsInStore } from '@/object-record/record-store/hooks/useUpsertRecordsInStore';
 import { useCreateNewIndexRecord } from '@/object-record/record-table/hooks/useCreateNewIndexRecord';
 import { Dropdown } from '@/ui/layout/dropdown/components/Dropdown';
 import { useToggleDropdown } from '@/ui/layout/dropdown/hooks/useToggleDropdown';
-import { useRecoilComponentCallbackState } from '@/ui/utilities/state/component-state/hooks/useRecoilComponentCallbackState';
 import { useRecoilComponentFamilyValue } from '@/ui/utilities/state/component-state/hooks/useRecoilComponentFamilyValue';
 import { useRecoilComponentValue } from '@/ui/utilities/state/component-state/hooks/useRecoilComponentValue';
-import { getSnapshotValue } from '@/ui/utilities/state/utils/getSnapshotValue';
-import { useRecoilCallback } from 'recoil';
-import { findByProperty, isDefined } from 'twenty-shared/utils';
 import { Tag } from 'twenty-ui/components';
 import { IconDotsVertical, IconPlus } from 'twenty-ui/display';
 import { LightIconButton } from 'twenty-ui/input';
@@ -104,62 +97,21 @@ export const RecordBoardColumnHeader = () => {
       recordIndexAggregateDisplayValueForGroupValueComponentFamilyState,
       { groupValue: columnDefinition?.value ?? '' },
     );
-  const { upsertRecordsInStore } = useUpsertRecordsInStore();
-
-  const recordIndexRecordIdsByGroupCallbackState =
-    useRecoilComponentCallbackState(
-      recordIndexRecordIdsByGroupComponentFamilyState,
-    );
 
   const recordIndexAggregateDisplayLabel = useRecoilComponentValue(
     recordIndexAggregateDisplayLabelComponentState,
-  );
-
-  const recordGroupDefinitions = useRecoilComponentValue(
-    recordGroupDefinitionsComponentSelector,
   );
 
   const { toggleDropdown } = useToggleDropdown();
 
   const dropdownId = `record-board-column-dropdown-${columnDefinition.id}`;
 
-  const handleCreateNewRecordClick = useRecoilCallback(
-    ({ set, snapshot }) =>
-      async () => {
-        const createdRecord = await createNewIndexRecord({
-          position: 'first',
-          [selectFieldMetadataItem.name]: columnDefinition.value,
-        });
-
-        const recordGroup = recordGroupDefinitions.find(
-          findByProperty('value', createdRecord[selectFieldMetadataItem.name]),
-        );
-
-        if (isDefined(recordGroup)) {
-          const currentRecordIds = getSnapshotValue(
-            snapshot,
-            recordIndexRecordIdsByGroupCallbackState(recordGroup.id),
-          );
-
-          const newRecordIds = [createdRecord.id, ...currentRecordIds];
-
-          set(
-            recordIndexRecordIdsByGroupCallbackState(recordGroup.id),
-            newRecordIds,
-          );
-        }
-
-        upsertRecordsInStore([createdRecord]);
-      },
-    [
-      createNewIndexRecord,
-      columnDefinition,
-      recordGroupDefinitions,
-      recordIndexRecordIdsByGroupCallbackState,
-      upsertRecordsInStore,
-      selectFieldMetadataItem,
-    ],
-  );
+  const handleCreateNewRecordClick = async () => {
+    await createNewIndexRecord({
+      position: 'first',
+      [selectFieldMetadataItem.name]: columnDefinition.value,
+    });
+  };
 
   return (
     <StyledColumn>
