+++ packages/twenty-front/src/modules/object-record/record-table/hooks/useCreateNewIndexRecord.ts
@@ -2,17 +2,24 @@ import { useOpenRecordInCommandMenu } from '@/command-menu/hooks/useOpenRecordIn
 import { type ObjectMetadataItem } from '@/object-metadata/types/ObjectMetadataItem';
 import { getLabelIdentifierFieldMetadataItem } from '@/object-metadata/utils/getLabelIdentifierFieldMetadataItem';
 import { useCreateOneRecord } from '@/object-record/hooks/useCreateOneRecord';
+import { recordGroupDefinitionsComponentSelector } from '@/object-record/record-group/states/selectors/recordGroupDefinitionsComponentSelector';
+import { recordIndexGroupFieldMetadataItemComponentState } from '@/object-record/record-index/states/recordIndexGroupFieldMetadataComponentState';
 import { recordIndexOpenRecordInState } from '@/object-record/record-index/states/recordIndexOpenRecordInState';
+import { recordIndexRecordIdsByGroupComponentFamilyState } from '@/object-record/record-index/states/recordIndexRecordIdsByGroupComponentFamilyState';
+import { useUpsertRecordsInStore } from '@/object-record/record-store/hooks/useUpsertRecordsInStore';
 import { useBuildRecordInputFromFilters } from '@/object-record/record-table/hooks/useBuildRecordInputFromFilters';
 import { useRecordTitleCell } from '@/object-record/record-title-cell/hooks/useRecordTitleCell';
 import { RecordTitleCellContainerType } from '@/object-record/record-title-cell/types/RecordTitleCellContainerType';
 import { type ObjectRecord } from '@/object-record/types/ObjectRecord';
 import { canOpenObjectInSidePanel } from '@/object-record/utils/canOpenObjectInSidePanel';
 import { getRecordFieldInputInstanceId } from '@/object-record/utils/getRecordFieldInputId';
+import { useRecoilComponentCallbackState } from '@/ui/utilities/state/component-state/hooks/useRecoilComponentCallbackState';
+import { useRecoilComponentValue } from '@/ui/utilities/state/component-state/hooks/useRecoilComponentValue';
+import { getSnapshotValue } from '@/ui/utilities/state/utils/getSnapshotValue';
 import { ViewOpenRecordInType } from '@/views/types/ViewOpenRecordInType';
 import { useRecoilCallback } from 'recoil';
 import { AppPath } from 'twenty-shared/types';
-import { isDefined } from 'twenty-shared/utils';
+import { findByProperty, isDefined } from 'twenty-shared/utils';
 import { v4 } from 'uuid';
 import { useNavigateApp } from '~/hooks/useNavigateApp';
 
@@ -23,12 +30,27 @@ type UseCreateNewIndexRecordProps = {
 export const useCreateNewIndexRecord = ({
   objectMetadataItem,
 }: UseCreateNewIndexRecordProps) => {
+  const recordGroupDefinitions = useRecoilComponentValue(
+    recordGroupDefinitionsComponentSelector,
+  );
+
+  const recordIndexRecordIdsByGroupCallbackState =
+    useRecoilComponentCallbackState(
+      recordIndexRecordIdsByGroupComponentFamilyState,
+    );
+
+  const recordIndexGroupFieldMetadataItem = useRecoilComponentValue(
+    recordIndexGroupFieldMetadataItemComponentState,
+  );
+
   const { openRecordInCommandMenu } = useOpenRecordInCommandMenu();
 
   const { createOneRecord } = useCreateOneRecord({
     objectNameSingular: objectMetadataItem.nameSingular,
     shouldMatchRootQueryFilter: true,
   });
+
+  const { upsertRecordsInStore } = useUpsertRecordsInStore();
 
   const navigate = useNavigateApp();
 
@@ -39,7 +61,7 @@ export const useCreateNewIndexRecord = ({
   });
 
   const createNewIndexRecord = useRecoilCallback(
-    ({ snapshot }) =>
+    ({ snapshot, set }) =>
       async (recordInput?: Partial<ObjectRecord>) => {
         const recordId = v4();
         const recordInputFromFilters = buildRecordInputFromFilters();
@@ -85,6 +107,40 @@ export const useCreateNewIndexRecord = ({
           });
         }
 
+        if (isDefined(recordIndexGroupFieldMetadataItem)) {
+          const recordGroup = recordGroupDefinitions.find(
+            findByProperty(
+              'value',
+              createdRecord[recordIndexGroupFieldMetadataItem.name],
+            ),
+          );
+
+          if (isDefined(recordGroup)) {
+            const currentRecordIds = getSnapshotValue(
+              snapshot,
+              recordIndexRecordIdsByGroupCallbackState(recordGroup.id),
+            );
+
+            if (recordInput?.position === 'first') {
+              const newRecordIds = [createdRecord.id, ...currentRecordIds];
+
+              set(
+                recordIndexRecordIdsByGroupCallbackState(recordGroup.id),
+                newRecordIds,
+              );
+            } else {
+              const newRecordIds = [...currentRecordIds, createdRecord.id];
+
+              set(
+                recordIndexRecordIdsByGroupCallbackState(recordGroup.id),
+                newRecordIds,
+              );
+            }
+          }
+        }
+
+        upsertRecordsInStore([createdRecord]);
+
         return createdRecord;
       },
     [
@@ -94,6 +150,10 @@ export const useCreateNewIndexRecord = ({
       objectMetadataItem,
       openRecordInCommandMenu,
       openRecordTitleCell,
+      recordGroupDefinitions,
+      recordIndexGroupFieldMetadataItem,
+      recordIndexRecordIdsByGroupCallbackState,
+      upsertRecordsInStore,
     ],
   );
